improved version with using only 1 malloc

+++ src/bits_operations.c
@@ -58,7 +58,7 @@ int fLireBit(FileBit *f)
         {
             return EOF;
         }
-        f->stock = n;
+        f->stock = (uint8_t)n;
         f->nbBit = 8;
     }
     f->nbBit--;
